requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.0
asttokens==2.4.1
asyncer==0.0.8
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==3.0.1
decorator==5.1.1
dill==0.3.8
evaluate==0.4.3
executing==2.1.0
fastapi==0.115.0
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gradio==5.0.2
gradio_client==1.4.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
ipython==8.28.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
megfile==3.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
open3d==0.18.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
optimum==1.23.1
orjson==3.10.7
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyMatting==1.1.12
PyMCubes==0.1.6
PyNaCl==1.5.0
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
rembg==2.0.59
requests==2.32.3
retrying==1.3.4
rich==13.9.2
rpds-py==0.20.0
ruff==0.6.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
trimesh==4.4.9
triton==3.0.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.1
watchdog==5.0.3
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
xformers==0.0.28.post1
xxhash==3.5.0
yarl==1.15.0
zipp==3.20.2
